+++ src/main/java/com/laundry_m/mvc/service/LaundryServiceImpl.java
@@ -105,13 +105,13 @@ public class LaundryServiceImpl implements LaundryService {
 		return laundryDao.selectByAddressLaundry(LaundryAddress);
 	}
 
-	public Laundry selectByLowestByLaundry(Long clothesId, Long FabricId)
+	public Laundry selectByLowestByLaundry(int clothesId, int FabricId)
 			throws SQLException, NotExistException, NotLoginException {
 		return laundryDao.selectByLowestByLaundry(clothesId, FabricId);
 	}
 
 	@Override
-	public int clothesfabricFee(Long laundryId, Long clothesId, Long fabricId)
+	public int clothesfabricFee(Long laundryId, int clothesId, int fabricId)
 			throws SQLException, NotExistException, NotLoginException {
 		return laundryDao.clothesfabricFee(laundryId, clothesId, fabricId);
 	}
